# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Fitness Coach API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  name?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface UserResponse {
  id: string;
  email: string;
  name?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserResponse;
}

export type UserProfileActivityLevel =
  (typeof UserProfileActivityLevel)[keyof typeof UserProfileActivityLevel];

export const UserProfileActivityLevel = {
  sedentary: "sedentary",
  lightly_active: "lightly_active",
  moderately_active: "moderately_active",
  very_active: "very_active",
  extremely_active: "extremely_active",
} as const;

export interface UserProfile {
  /** Height (e.g. "5ft 10in" or "178cm") */
  height: string;
  /** Weight (e.g. "75kg" or "165lbs") */
  weight: string;
  age: number;
  activityLevel: UserProfileActivityLevel;
  /** Fitness goals */
  goals: string;
  /** Any injuries or physical limitations */
  injuries?: string | null;
  /** Dietary allergies or restrictions */
  allergies?: string | null;
  /** Available workout equipment (e.g. full gym, dumbbells only, no equipment) */
  equipment?: string | null;
}

export interface GenerateWorkoutRequest {
  userProfile: UserProfile;
  /** Base64-encoded image of user physique (optional) */
  photoBase64?: string | null;
}

/**
 * The AI-generated workout plan
 */
export type WorkoutPlanResponsePlan = { [key: string]: unknown };

export interface WorkoutPlanResponse {
  id: string;
  userId: string;
  /** The AI-generated workout plan */
  plan: WorkoutPlanResponsePlan;
  userProfile: UserProfile;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface CreateProgressRequest {
  /** The ID of the workout plan this progress belongs to */
  planId: string;
  /** Base64-encoded image of daily progress */
  photoBase64: string;
}

export interface ProgressEntryResponse {
  id: string;
  userId: string;
  planId: string;
  photoBase64: string;
  /** AI-generated feedback for the progress */
  aiFeedback: string;
  createdAt: string;
}

export type GetProgressEntriesParams = {
  /**
   * Filter progress entries by workout plan ID
   */
  planId: string;
};