# lib/api-zod/src/generated/types/progressEntryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Fitness Coach API
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressEntryResponse {
  id: string;
  userId: string;
  planId: string;
  photoBase64: string;
  /** AI-generated feedback for the progress */
  aiFeedback: string;
  createdAt: Date;
}