# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Fitness Coach API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileActivityLevel } from "./userProfileActivityLevel";

export interface UserProfile {
  /** Height (e.g. "5ft 10in" or "178cm") */
  height: string;
  /** Weight (e.g. "75kg" or "165lbs") */
  weight: string;
  age: number;
  activityLevel: UserProfileActivityLevel;
  /** Fitness goals */
  goals: string;
  /** Any injuries or physical limitations */
  injuries?: string | null;
  /** Dietary allergies or restrictions */
  allergies?: string | null;
  /** Available workout equipment (e.g. full gym, dumbbells only, no equipment) */
  equipment?: string | null;
}